Remove API routes and switch to server actions
